move user controller to core/controller
